Simplfy and clarify plugin function signature
This commit simplifies the handle_files signature to
take in, as a parameter, the GitPuller reference from nbgitpuller
as well as renames the handle_files function signature to
prepare_non_git_source_local_origin for clarity.

Due to changes in the nbgitpuller plugin implementation we can now
simply return the json with the appropriate information from
the prepare_non_git_source_local_origin function.

The tests and documentation are updated to reflect these changes.